Lab1_DockerForDatabases/app/requirements.txt
flask==2.1.0
flask_sqlalchemy==2.5.1
mysqlclient==2.0.3